requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
arrow==1.2.2
asgiref==3.2.10
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.10.0
blessed==1.19.1
certifi==2021.10.8
charset-normalizer==2.0.12
dateparser==1.1.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.2
django-crontab==0.7.1
django-extensions==3.1.5
django-mathfilters==1.0.0
django-picklefield==3.0.1
django-q==1.3.9
frozenlist==1.3.0
future==0.18.2
GoogleNews==1.6.0
graphviz==0.19.1
gunicorn==20.0.4
idna==3.3
line-bot-sdk==2.2.1
lxml==4.8.0
multidict==6.0.2
multitasking==0.0.10
numpy==1.22.3
pandas==1.4.1
psycopg2-binary==2.8.6
pyparsing==3.0.7
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==3.5.3
regex==2022.3.2
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
static3==0.7.0
twstock==1.3.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
wcwidth==0.2.5
websockets==9.1
whitenoise==5.2.0
yarl==1.7.2
yfinance==0.1.70
